requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
# asyncpg==0.30.0  # production only (PostgreSQL) — requires C++ build tools on Windows
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
azure-ai-formrecognizer==3.3.3
azure-identity==1.19.0
httpx==0.27.0
pydantic-settings==2.6.0
alembic==1.13.3
aiosqlite==0.20.0
aiofiles==23.2.1
# gunicorn==22.0.0  # production only — not available on Windows
pytest==8.3.0
pytest-asyncio==0.24.0
